+++ crates/hyperswitch_connectors/src/types.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "v2")]
+use hyperswitch_domain_models::router_data_v2::RouterDataV2;
 #[cfg(feature = "payouts")]
 use hyperswitch_domain_models::types::{PayoutsData, PayoutsResponseData};
 use hyperswitch_domain_models::{
@@ -6,16 +8,13 @@ use hyperswitch_domain_models::{
         authentication::{
             Authentication, PostAuthentication, PreAuthentication, PreAuthenticationVersionCall,
         },
-        Accept, AccessTokenAuth, Authorize, Capture, Checkout, Defend, Evidence, Fulfillment,
-        PSync, PostProcessing, PreProcessing, Session, Transaction, Upload, Void,
+        Accept, AccessTokenAuth, Authorize, Capture, Defend, Evidence, PSync, PostProcessing,
+        PreProcessing, Session, Upload, Void,
     },
     router_request_types::{
         authentication::{
             ConnectorAuthenticationRequestData, ConnectorPostAuthenticationRequestData,
             PreAuthNRequestData,
-        },
-        fraud_check::{
-            FraudCheckCheckoutData, FraudCheckFulfillmentData, FraudCheckTransactionData,
         },
         AcceptDisputeRequestData, AccessTokenRequestData, DefendDisputeRequestData,
         PaymentsAuthorizeData, PaymentsCancelData, PaymentsCaptureData, PaymentsPostProcessingData,
@@ -23,15 +22,18 @@ use hyperswitch_domain_models::{
         SubmitEvidenceRequestData, UploadFileRequestData,
     },
     router_response_types::{
-        fraud_check::FraudCheckResponseData, AcceptDisputeResponse, AuthenticationResponseData,
-        DefendDisputeResponse, PaymentsResponseData, RefundsResponseData, SubmitEvidenceResponse,
-        UploadFileResponse,
+        AcceptDisputeResponse, AuthenticationResponseData, DefendDisputeResponse,
+        PaymentsResponseData, RefundsResponseData, SubmitEvidenceResponse, UploadFileResponse,
     },
 };
 #[cfg(feature = "frm")]
 use hyperswitch_domain_models::{
-    router_flow_types::{RecordReturn, Sale},
-    router_request_types::fraud_check::{FraudCheckRecordReturnData, FraudCheckSaleData},
+    router_flow_types::{Checkout, Fulfillment, RecordReturn, Sale, Transaction},
+    router_request_types::fraud_check::{
+        FraudCheckCheckoutData, FraudCheckFulfillmentData, FraudCheckRecordReturnData,
+        FraudCheckSaleData, FraudCheckTransactionData,
+    },
+    router_response_types::fraud_check::FraudCheckResponseData,
 };
 use hyperswitch_interfaces::api::ConnectorIntegration;
 
@@ -63,58 +65,70 @@ pub(crate) type DefendDisputeRouterData =
     RouterData<Defend, DefendDisputeRequestData, DefendDisputeResponse>;
 
 #[cfg(feature = "payouts")]
-pub type PayoutsResponseRouterData<F, R> =
+pub(crate) type PayoutsResponseRouterData<F, R> =
     ResponseRouterData<F, R, PayoutsData, PayoutsResponseData>;
 
 // TODO: Remove `ResponseRouterData` from router crate after all the related type aliases are moved to this crate.
-pub struct ResponseRouterData<Flow, R, Request, Response> {
+pub(crate) struct ResponseRouterData<Flow, R, Request, Response> {
+    pub(crate) response: R,
+    pub(crate) data: RouterData<Flow, Request, Response>,
+    pub(crate) http_code: u16,
+}
+#[cfg(feature = "frm")]
+pub(crate) type FrmFulfillmentRouterData =
+    RouterData<Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>;
+#[cfg(feature = "frm")]
+pub(crate) type FrmCheckoutType =
+    dyn ConnectorIntegration<Checkout, FraudCheckCheckoutData, FraudCheckResponseData>;
+#[cfg(feature = "frm")]
+pub(crate) type FrmTransactionType =
+    dyn ConnectorIntegration<Transaction, FraudCheckTransactionData, FraudCheckResponseData>;
+#[cfg(feature = "frm")]
+pub(crate) type FrmTransactionRouterData =
+    RouterData<Transaction, FraudCheckTransactionData, FraudCheckResponseData>;
+#[cfg(feature = "frm")]
+pub(crate) type FrmFulfillmentType =
+    dyn ConnectorIntegration<Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>;
+#[cfg(feature = "frm")]
+pub(crate) type FrmCheckoutRouterData =
+    RouterData<Checkout, FraudCheckCheckoutData, FraudCheckResponseData>;
+#[cfg(feature = "v2")]
+pub(crate) struct ResponseRouterDataV2<Flow, R, ResourceCommonData, Request, Response> {
     pub response: R,
-    pub data: RouterData<Flow, Request, Response>,
+    pub data: RouterDataV2<Flow, ResourceCommonData, Request, Response>,
     pub http_code: u16,
 }
-pub type FrmFulfillmentRouterData =
-    RouterData<Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>;
-pub type FrmCheckoutType =
-    dyn ConnectorIntegration<Checkout, FraudCheckCheckoutData, FraudCheckResponseData>;
-pub type FrmTransactionType =
-    dyn ConnectorIntegration<Transaction, FraudCheckTransactionData, FraudCheckResponseData>;
-pub type FrmTransactionRouterData =
-    RouterData<Transaction, FraudCheckTransactionData, FraudCheckResponseData>;
-pub type FrmFulfillmentType =
-    dyn ConnectorIntegration<Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>;
-pub type FrmCheckoutRouterData =
-    RouterData<Checkout, FraudCheckCheckoutData, FraudCheckResponseData>;
 
-pub type PreAuthNRouterData =
+pub(crate) type PreAuthNRouterData =
     RouterData<PreAuthentication, PreAuthNRequestData, AuthenticationResponseData>;
-pub type PreAuthNVersionCallRouterData =
+pub(crate) type PreAuthNVersionCallRouterData =
     RouterData<PreAuthenticationVersionCall, PreAuthNRequestData, AuthenticationResponseData>;
-pub type ConnectorAuthenticationRouterData =
+pub(crate) type ConnectorAuthenticationRouterData =
     RouterData<Authentication, ConnectorAuthenticationRequestData, AuthenticationResponseData>;
-pub type ConnectorPostAuthenticationRouterData = RouterData<
+pub(crate) type ConnectorPostAuthenticationRouterData = RouterData<
     PostAuthentication,
     ConnectorPostAuthenticationRequestData,
     AuthenticationResponseData,
 >;
-pub type ConnectorAuthenticationType = dyn ConnectorIntegration<
+pub(crate) type ConnectorAuthenticationType = dyn ConnectorIntegration<
     Authentication,
     ConnectorAuthenticationRequestData,
     AuthenticationResponseData,
 >;
-pub type ConnectorPostAuthenticationType = dyn ConnectorIntegration<
+pub(crate) type ConnectorPostAuthenticationType = dyn ConnectorIntegration<
     PostAuthentication,
     ConnectorPostAuthenticationRequestData,
     AuthenticationResponseData,
 >;
-pub type ConnectorPreAuthenticationType =
+pub(crate) type ConnectorPreAuthenticationType =
     dyn ConnectorIntegration<PreAuthentication, PreAuthNRequestData, AuthenticationResponseData>;
-pub type ConnectorPreAuthenticationVersionCallType = dyn ConnectorIntegration<
+pub(crate) type ConnectorPreAuthenticationVersionCallType = dyn ConnectorIntegration<
     PreAuthenticationVersionCall,
     PreAuthNRequestData,
     AuthenticationResponseData,
 >;
 
-pub type PaymentsPostProcessingRouterData =
+pub(crate) type PaymentsPostProcessingRouterData =
     RouterData<PostProcessing, PaymentsPostProcessingData, PaymentsResponseData>;
 #[cfg(feature = "frm")]
 pub(crate) type FrmSaleRouterData = RouterData<Sale, FraudCheckSaleData, FraudCheckResponseData>;